Move old query model into `query_engines`
It should now be considered an implementation detail of the Query
Engine, rather than a module that anything else should depend on.
