+++ databuilder/__init__.py
@@ -2,7 +2,7 @@ from .codelistlib import codelist, codelist_from_csv, combine_codelists
 from .date_utils import cohort_date_range
 from .log_utils import init_logging
 from .measure import Measure
-from .query_model_old import categorise, table
+from .query_engines.query_model_old import categorise, table
 
 init_logging()
