+++ tests/test_query_model_old.py
@@ -1,7 +1,7 @@
 import pytest
 
 from databuilder.codelistlib import Codelist
-from databuilder.query_model_old import (
+from databuilder.query_engines.query_model_old import (
     Comparator,
     FilteredTable,
     Row,
